+++ src/Widgets/NodeEditor/NodeIOConnector.cpp
@@ -3,12 +3,12 @@
 #include "Registry.hpp"
 #include "Widgets/NodeEditor.hpp"
 
-NodeIOConnector::NodeIOConnector(NodeInputGraphics* parent, float x, float y) : QWidget(parent)
+NodeIOConnector::NodeIOConnector(NodeInputGraphics* parent, int x, int y) : QWidget(parent)
 {
     init(x, y);
 }
 
-NodeIOConnector::NodeIOConnector(NodeOutputGraphics* parent, float x, float y) : QWidget(parent)
+NodeIOConnector::NodeIOConnector(NodeOutputGraphics* parent, int x, int y) : QWidget(parent)
 {
     init(x, y);
 }
@@ -34,13 +34,13 @@ void NodeIOConnector::paintEvent(QPaintEvent*)
     painter.setRenderHint(QPainter::Antialiasing);
     QPainterPath border{};
     border.addEllipse(0, 0, Registry::getRegistry()->extrinsic->GUI->dimensions["NodeConnectorDiameter"], Registry::getRegistry()->extrinsic->GUI->dimensions["NodeConnectorDiameter"]);
-    QPen borderPen{Registry::getRegistry()->extrinsic->GUI->palette["NodeConnectorBorder"], Registry::getRegistry()->extrinsic->GUI->dimensions["NodeConnectorBorder"]};
+    QPen borderPen{Registry::getRegistry()->extrinsic->GUI->palette["NodeConnectorBorder"], static_cast<double>(Registry::getRegistry()->extrinsic->GUI->dimensions["NodeConnectorBorder"])};
     painter.setPen(borderPen);
     painter.fillPath(border, Registry::getRegistry()->extrinsic->GUI->palette["NodeConnectorBackground"]);
     painter.drawPath(border);
 }
 
-void NodeIOConnector::init(float x, float y)
+void NodeIOConnector::init(int x, int y)
 {
     move(x, y);
     resize(Registry::getRegistry()->extrinsic->GUI->dimensions["NodeConnectorDiameter"], Registry::getRegistry()->extrinsic->GUI->dimensions["NodeConnectorDiameter"]);
